python-server/requirements.txt
Click==7.0
decorator==4.4.0
Flask==1.0.3
Flask-Cors==3.0.7
imageio==2.5.0
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
networkx==2.3
numpy==1.16.4
Pillow==6.0.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
Werkzeug==0.15.4
